Unify common mongod and mongos server metrics
New collector_common package for metrics that are the same in both
collector_mongod and collector_mongos packages.

The labels of these metrics are now prefixed with just 'mongodb_'
rather than 'mongodb_mongo[ds]_'.

Note that because bson only supports ",inline" for maps and structs,
the common ServerStatus is embedded by-value not by-pointer.