Expose exporter.Handler and allow disabling default registry
The Grafana Agent (https://github.com/grafana/agent) embeds
mongodb_exporter as an integration, where it needs to be able to call
into the exporter's HTTP handler for its all-in-one server. This
endpoint also needs to only return metrics for the exporter itself and
not from the global registry, which is populated by unrelated metrics.

Signed-off-by: Robert Fratto <robertfratto@gmail.com>