Enable Basic Auth on new HTTP server
This is intended to be applied on top of #393 (the re-revert) and adds a
new parameter, the tls-config-path file that should be sent by
pmm-managed, pointing to a file containing the credentials we use to
enable basic auth.

In the original implementation, we use exporter_shared library which
receives basic auth credentials via env vars.
Now, using Prometheus Toolkit library, we need to pass a file.