collstats: fix compatibility issue and disable latency histograms (#433)
$collStats is invoked by exporter to get collection statistics from
MongoDB. The configuration generated by exporter worked for v4.x
of MongoDB, but failed for v5.0. This change modifies configuration
 to work for v4.x and v5.0.

Once fix works, latency statistics are generated in form of histogram.
MongoDB exporter doesn't generate histogram metrics for Prometheus
as of yet. Hence we disable histogram for latency statistics in this patch.