PMM-9757 Fixed panic on lost connection
Changed the way we instantiate the topology info object.
Now we call the newTopologyInfo and it never returns an error and the
object is never nill. If it is not possible to connect, topology labels
remain empty but we return a valid and not nill object.
Also added context termination check to avoid the ugly stack traces
produced by context timeouts.