PMM-10271 Fix meta-metric label for general collector
Meta-metric generated for general collector were incorrectly labeled as
dbstats. When dbstats and general collector were executed for a scrape request,
two metric were generated with the same name and labels. This duplicate metric
would trigger a panic in the metric generation library. This change updates the
label for general collector and fixes the issue.