PMM-14141: Respect Prometheus scrape-timeout header (#1107)
* Fix parsing of X-Prometheus-Scrape-Timeout-Seconds header
The previous implementation expected this header to be an integer, but vmagent
 may send it as a float. This caused incorrect handling in `mongodb_exporter`.

Updated to support float values.

* Avoid stale connections by pinging on acquire
The exporter was returning cached MongoDB connections without validating their health. This meant that if a connection had become unhealthy,
subsequent operations would fail after a delay (serverSelectionTimeout etc), causing the entire request to be slower than it should be.

This commit adds  `Ping` check before returning a cached client.

* Update exporter/exporter.go

Co-authored-by: Alex Demidoff <alexander.demidoff@percona.com>

---------

Co-authored-by: Alex Demidoff <alexander.demidoff@percona.com>
Co-authored-by: Michael Okoko <10512379+idoqo@users.noreply.github.com>